requirements.txt
openai
transformers
python-dotenv
click
torch
datasets
